+++ web/rainmaker/dev-packages/egov-pt-dev/src/ui-config/screens/specs/pt-mutation/applyResourceMutation/mutationDetails.js
@@ -20,125 +20,6 @@ import {
   import { handleScreenConfigurationFieldChange as handleField } from "egov-ui-framework/ui-redux/screen-configuration/actions";
   import get from "lodash/get";
 
-  export const transferorDetails = getCommonCard(
-    {
-      header: getCommonTitle(
-        {
-          labelName: "Transferor Details",
-          labelKey: "PT_MUTATION_TRANSFEROR_DETAILS"
-        },
-        {
-          style: {
-            marginBottom: 18
-          }
-        }
-      ),
-      body: getCommonContainer({
-        transferorName: getLabelWithValue(
-          {
-            labelName: "Name",
-            labelKey: "PT_MUTATION_TRANSFEROR_NAME"
-          },
-          {
-            jsonPath: "FireNOCs[0].fireNOCDetails.fireNOCType"
-            // callBack: value => {
-            //   return value.split(".")[0];
-            // }
-          }
-        ),
-        guardianName: getLabelWithValue(
-          {
-            labelName: "Guardian's Name",
-            labelKey: "PT_MUTATION_TRANSFEROR_GUARDIAN_NAME"
-          },
-          {
-            jsonPath: "FireNOCs[0].provisionFireNOCNumber"
-            // callBack: value => {
-            //   return value.split(".")[1];
-            // }
-          }
-        ),
-        transferorGender: getLabelWithValue(
-          {
-            labelName: "Gender",
-            labelKey: "PT_MUTATION_TRANSFEROR_GENDER"
-          },
-          {
-            jsonPath: "FireNOCs[0].provisionFireNOCNumber"
-            // callBack: value => {
-            //   return value.split(".")[1];
-            // }
-          }
-        ),
-        transferorDOB: getLabelWithValue(
-          {
-            labelName: "Date Of Birth",
-            labelKey: "PT_MUTATION_TRANSFEROR_DOB"
-          },
-          {
-            jsonPath: "FireNOCs[0].provisionFireNOCNumber"
-            // callBack: value => {
-            //   return value.split(".")[1];
-            // }
-          }
-        ),
-        transferorMobile: getLabelWithValue(
-          {
-            labelName: "Mobile No.",
-            labelKey: "PT_MUTATION_TRANSFEROR_MOBILE"
-          },
-          {
-            jsonPath: "FireNOCs[0].provisionFireNOCNumber"
-            // callBack: value => {
-            //   return value.split(".")[1];
-            // }
-          }
-        ),
-        transferorEmail: getLabelWithValue(
-          {
-            labelName: "Email",
-            labelKey: "PT_MUTATION_TRANSFEROR_EMAIL"
-          },
-          {
-            jsonPath: "FireNOCs[0].provisionFireNOCNumber"
-            // callBack: value => {
-            //   return value.split(".")[1];
-            // }
-          }
-        ),
-        transferorSpecialCategory: getLabelWithValue(
-          {
-            labelName: "Special Category",
-            labelKey: "PT_MUTATION_TRANSFEROR_SPECIAL_CATEGORY"
-          },
-          {
-            jsonPath: "FireNOCs[0].provisionFireNOCNumber"
-            // callBack: value => {
-            //   return value.split(".")[1];
-            // }
-          }
-        ),
-        transferorCorrespondenceAddress: getLabelWithValue(
-          {
-            labelName: "Correspondence Address",
-            labelKey: "PT_MUTATION_TRANSFEROR_CORRESPONDENCE_ADDRESS"
-          },
-          {
-            jsonPath: "FireNOCs[0].provisionFireNOCNumber"
-            // callBack: value => {
-            //   return value.split(".")[1];
-            // }
-          }
-        )
-      })
-    });
-
-    
-
-
-
-
-
 
       export const mutationDetails = getCommonCard(
         {
@@ -254,102 +135,7 @@ import {
           })
         });
 
-        export const registrationDetails = getCommonCard(
-          {
-            header: getCommonTitle(
-              {
-                labelName: "Registration Details",
-                labelKey: "PT_MUTATION_REGISTRATION_DETAILS"
-              },
-              {
-                style: {
-                  marginBottom: 18
-                }
-              }
-            ),
-            registrationDetailsContainer: getCommonContainer({ 
-              transferReason: getSelectField({
-                label: {
-                  labelName: "Reason For Transfer",
-                  labelKey: "PT_MUTATION_TRANSFER_REASON"
-                },
-                placeholder: {
-                  labelName: "Select Reason for Transfer",
-                  labelKey: "PT_MUTATION_TRANSFER_REASON_PLACEHOLDER"
-                },
-                required:true,
-                jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].ownerType",
-                sourceJsonPath: "applyScreenMdmsData.common-masters.OwnerType",
-                localePrefix: {
-                  moduleName: "common-masters",
-                  masterName: "OwnerType"
-                }
-              }),
-              documentNumber: getTextField({
-                label: {
-                  labelName: "Document No.",
-                  labelKey: "PT_MUTATION_DOCUMENT_NO"
-                },
-                props:{
-                  className:"applicant-details-error"
-                },
-                placeholder: {
-                  labelName: "Enter Document No.",
-                  labelKey: "PT_MUTATION_DOCUMENT_NO_PLACEHOLDER"
-                },
-                required:true,
-                pattern: getPattern("Address"),
-          //      jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
-              }),
-                documentIssueDateField :getDateField({
-                label: { labelName: "Document Issue Date", labelKey: "PT_MUTATION_DOCUMENT_ISSUE_DATE" },
-                placeholder: {
-                  labelName: "Enter Document No.",
-                  labelKey: "PT_MUTATION_DOCUMENT_ISSUE_DATE_PLACEHOLDER"
-                },
-                required: true,
-                pattern: getPattern("Date"),
-                jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].dob",
-                // props: {
-                //   inputProps: {
-                //     max: getTodaysDateInYMD()
-                //   }
-                // }
-              }),
-              documentValue: getTextField({
-                label: {
-                  labelName: "Document Value",
-                  labelKey: "PT_MUTATION_DOCUMENT_VALUE"
-                },
-                props:{
-                  className:"applicant-details-error"
-                },
-                placeholder: {
-                  labelName: "Enter Document Value",
-                  labelKey: "PT_MUTATION_DOCUMENT_VALUE_PLACEHOLDER"
-                },
-                required:true,
-                pattern: getPattern("Address"),
-           //     jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
-              }),
-              remarks: getTextField({
-                label: {
-                  labelName: "Remarks",
-                  labelKey: "PT_MUTATION_REMARKS"
-                },
-                props:{
-                  className:"applicant-details-error"
-                },
-                placeholder: {
-                  labelName: "Enter Remarks if any",
-                  labelKey: "PT_MUTATION_REMARKS"
-                },
-                pattern: getPattern("Address"),
-               // jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
-              }),
-            })
-          }) ; 
-
+
