+++ web/rainmaker/dev-packages/egov-pt-dev/src/ui-config/screens/specs/pt-mutation/applyResourceMutation/transfereeDetails.js
@@ -187,41 +187,36 @@ const commonApplicantInformation = () => {
           className:"applicant-details-error"
         }
       }),
-      relationshipRadioGroup: {
-        uiFramework: "custom-containers",
-        componentPath: "RadioGroupContainer",
+      relationshipWithGuardian: getSelectField({
+        label: {
+          labelName: "Relationship with Guardian",
+          labelKey: "PT_MUTATION_TRANSFEREE_APPLICANT_RELATIONSHIP_LABEL"
+        },
+        placeholder: {
+          labelName: "Select Relationship with Guardian",
+          labelKey: "PT_MUTATION_TRANSFEREE_APPLICANT_RELATIONSHIP_LABEL_PLACEHOLDER"
+        },
+        jsonPath:
+          "FireNOCs[0].fireNOCDetails.applicantDetails.owners[0].ownerType",
+        data: [
+          {
+            code: "A"
+          },
+          {
+            code: "B"
+          }
+        ],
+        localePrefix: {
+          moduleName: "common-masters",
+          masterName: "OwnerType"
+        },
+        //sourceJsonPath: "applyScreenMdmsData.common-masters.OwnerType",
         gridDefination: {
           xs: 12,
           sm: 12,
           md: 6
-        },
-        jsonPath:
-          "FireNOCs[0].fireNOCDetails.applicantDetails.owners[0].relationship",
-        props: {
-          label: {
-            name: "Relationship",
-            key: "PT_MUTATION_TRANSFEREE_APPLICANT_RELATIONSHIP_LABEL",
-            className:"applicant-details-error"
-          },
-          buttons: [
-            {
-              labelName: "Father",
-              labelKey: "PT_MUTATION_TRANSFEREE_APPLICANT_RELATIONSHIP_FATHER_RADIOBUTTON",
-              value: "FATHER"
-            },
-            {
-              label: "Husband",
-              labelKey: "PT_MUTATION_TRANSFEREE_APPLICANT_RELATIONSHIP_HUSBAND_RADIOBUTTON",
-              value: "HUSBAND"
-            }
-          ],
-          jsonPath:
-            "FireNOCs[0].fireNOCDetails.applicantDetails.owners[0].relationship",
-          required: true
-        },
-        required: true,
-        type: "array"
-      },
+        }
+      }),
       
       applicantAddress: getTextField({
         label: {
@@ -281,152 +276,141 @@ const commonApplicantInformation = () => {
 };
 
 const institutionInformation = () => {
-  return getCommonGrayCard({
-   
-    applicantCard: getCommonContainer({
-      institutionName: getTextField({
-        label: {
-          labelName: "Name of Institution",
-          labelKey: "PT_MUTATION_TRANSFEREE_INSTITUTION_LABEL"
+  return getCommonCard(
+    {
+      header: getCommonTitle(
+        {
+          labelName: "Details of Authorised Person",
+          labelKey: "PT_MUTATION_AUTHORISED_PERSON_DETAILS"
         },
-        placeholder: {
-          labelName: "Enter Name of Institution",
-          labelKey: "PT_MUTATION_TRANSFEREE_INSTITUTION_PLACEHOLDER"
-        },
-        pattern: getPattern("Name"),
-        errorMessage: "Invalid Name",
-        required: true,
-        jsonPath:
-          "FireNOCs[0].fireNOCDetails.applicantDetails.additionalDetail.institutionName",
-        gridDefination: {
-          xs: 12,
-          sm: 12,
-          md: 6
+        {
+          style: {
+            marginBottom: 18
+          }
         }
+      ),
+      institutionTypeDetailsContainer: getCommonContainer({ 
+        
+        privateInstitutionNameDetails: getTextField({
+          label: {
+            labelName: "Name",
+            labelKey: "PT_MUTATION_INSTITUTION_NAME"
+          },
+          props:{
+            className:"applicant-details-error"
+          },
+          placeholder: {
+            labelName: "Enter Name",
+            labelKey: "PT_MUTATION_INSTITUTION_NAME_PLACEHOLDER"
+          },
+          required:true,
+          pattern: getPattern("Name"),
+    //      jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
+        }),
       }),
-      telephoneNumber: getTextField({
-        label: {
-          labelName: "Official Telephone No.",
-          labelKey: "PT_MUTATION_TRANSFEREE_TELEPHONE_NUMBER_LABEL"
-        },
-        placeholder: {
-          labelName: "Enter Official Telephone No.",
-          labelKey: "PT_MUTATION_TRANSFEREE_TELEPHONE_NUMBER_PLACEHOLDER"
-        },
-        required: true,
-        pattern: getPattern("MobileNo"),
-        errorMessage: "Invalid Number",
-        jsonPath:
-          "FireNOCs[0].fireNOCDetails.applicantDetails.additionalDetail.telephoneNumber",
-        gridDefination: {
-          xs: 12,
-          sm: 12,
-          md: 6
-        }
-      }),
-      authorisedPerson: getTextField({
-        label: {
-          labelName: "Name of Authorized Person",
-          labelKey: "PT_MUTATION_TRANSFEREE_AUTHORIZED_PERSON_LABEL"
-        },
-        placeholder: {
-          labelName: "Enter Name of Authorized Person",
-          labelKey: "PT_MUTATION_TRANSFEREE_AUTHORIZED_PERSON_PLACEHOLDER"
-        },
-        required: true,
-        pattern: getPattern("Name"),
-        errorMessage: "Invalid Name",
-        jsonPath: "FireNOCs[0].fireNOCDetails.applicantDetails.owners[0].name",
-        gridDefination: {
-          xs: 12,
-          sm: 12,
-          md: 6
-        }
-      }),
-      designation: getTextField({
-        label: {
-          labelName: "Designation in Institution",
-          labelKey: "PT_MUTATION_TRANSFEREE_INSTITUTION_DESIGNATION_LABEL"
-        },
-        placeholder: {
-          labelName: "Enter designation of Institution",
-          labelKey: "PT_MUTATION_TRANSFEREE_INSTITUTION_DESIGNATION_PLACEHOLDER"
-        },
-        required: true,
-        pattern: getPattern("Name"),
-        errorMessage: "Invalid Designation Name",
-        jsonPath:
-          "FireNOCs[0].fireNOCDetails.applicantDetails.additionalDetail.institutionDesignation",
-        gridDefination: {
-          xs: 12,
-          sm: 12,
-          md: 6
-        }
-      }),
-      authorizedPersonMobile: getTextField({
-        label: {
-          labelName: "Mobile No. of Authorized Person",
-          labelKey: "PT_MUTATION_TRANSFEREE_AUTHORIZED_PERSON_MOBILE_LABEL"
-        },
-        placeholder: {
-          labelName: "Enter Mobile No. of Authorized Person",
-          labelKey: "PT_MUTATION_TRANSFEREE_AUTHORIZED_PERSON_MOBILE_PLACEHOLDER"
-        },
-        required: true,
-        pattern: getPattern("MobileNo"),
-        errorMessage: "Invalid MobileNo.",
 
-        jsonPath:
-          "FireNOCs[0].fireNOCDetails.applicantDetails.owners[0].mobileNumber",
-        gridDefination: {
-          xs: 12,
-          sm: 12,
-          md: 6
-        }
-      }),
-      authorizedPersonEmail: getTextField({
-        label: {
-          labelName: "Email of Authorized Person",
-          labelKey: "PT_MUTATION_AUTHORIZED_PERSON_EMAIL_LABEL"
-        },
-        placeholder: {
-          labelName: "Enter Email of Authorized Person",
-          labelKey: "PT_MUTATION_AUTHORIZED_PERSON_EMAIL_PLACEHOLDER"
-        },
-        pattern: getPattern("Email"),
-        errorMessage: "Invalid Email",
-        required: true,
-        jsonPath:
-          "FireNOCs[0].fireNOCDetails.applicantDetails.owners[0].emailId",
-        gridDefination: {
-          xs: 12,
-          sm: 12,
-          md: 6
-        }
-      }),
-      officialCorrespondenceAddress: getTextField({
-        label: {
-          labelName: "Official Correspondence Address",
-          labelKey: "PT_MUTATION_CORRESPONDENCE_ADDRESS_LABEL"
-        },
-        placeholder: {
-          labelName: "Enter Official Correspondence Address ",
-          labelKey: "PT_MUTATION_OFFICIAL_CORRESPONDENCE_ADDRESS_PLACEHOLDER"
-        },
-        required: true,
-        pattern: getPattern("Address"),
-        errorMessage: "Invalid Address",
-        jsonPath:
-          "FireNOCs[0].fireNOCDetails.applicantDetails.owners[0].correspondenceAddress",
-        gridDefination: {
-          xs: 12,
-          sm: 12,
-          md: 6
-        }
+      institutionDetailsContainer: getCommonContainer({ 
+        
+        authorisedPersonName: getTextField({
+          label: {
+            labelName: "Name",
+            labelKey: "PT_MUTATION_AUTHORISED_PERSON_NAME"
+          },
+          props:{
+            className:"applicant-details-error"
+          },
+          placeholder: {
+            labelName: "Enter Name",
+            labelKey: "PT_MUTATION_AUTHORISED_PERSON_NAME_PLACEHOLDER"
+          },
+          required:true,
+          pattern: getPattern("Name"),
+    //      jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
+        }),
+          
+        authorisedDesignationValue: getTextField({
+          label: {
+            labelName: "Designation",
+            labelKey: "PT_MUTATION_AUTHORISED_PERSON_DESIGNATION"
+          },
+          props:{
+            className:"applicant-details-error"
+          },
+          placeholder: {
+            labelName: "Enter Designation",
+            labelKey: "PT_MUTATION_AUTHORISED_PERSON_DESIGNATION_PLACEHOLDER"
+          },
+          required:true,
+          pattern: getPattern("Name"),
+     //     jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
+        }),
+        authorisedMobile: getTextField({
+          label: {
+            labelName: "Mobile",
+            labelKey: "PT_MUTATION_AUTHORISED_MOBILE"
+          },
+          props:{
+            className:"applicant-details-error"
+          },
+          placeholder: {
+            labelName: "Enter Mobile",
+            labelKey: "PT_MUTATION_AUTHORISED_MOBILE_PLACEHOLDER"
+          },
+          required:true,
+          pattern: getPattern("Number"),
+         // jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
+        }),
+        authorisedLandline: getTextField({
+            label: {
+              labelName: "Landline",
+              labelKey: "PT_MUTATION_AUTHORISED_LANDLINE"
+            },
+            props:{
+              className:"applicant-details-error"
+            },
+            placeholder: {
+              labelName: "Enter Landline",
+              labelKey: "PT_MUTATION_AUTHORISED_LANDLINE_PLACEHOLDER"
+            },
+            pattern: getPattern("MobileNo"),
+           // jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
+          }),
+          authorisedEmail: getTextField({
+            label: {
+              labelName: "Email",
+              labelKey: "PT_MUTATION_AUTHORISED_EMAIL"
+            },
+            props:{
+              className:"applicant-details-error"
+            },
+            placeholder: {
+              labelName: "Enter Email",
+              labelKey: "PT_MUTATION_AUTHORISED_EMAIL_PLACEHOLDER"
+            },
+            pattern: getPattern("Email"),
+           // jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
+          }),
+          authorisedAddress: getTextField({
+            label: {
+              labelName: "Correspondence Address",
+              labelKey: "PT_MUTATION_AUTHORISED_CORRESPONDENCE_ADDRESS"
+            },
+            props:{
+              className:"applicant-details-error"
+            },
+            placeholder: {
+              labelName: "Enter Correspondence Address",
+              labelKey: "PT_MUTATION_AUTHORISED_ADDRESS_PLACEHOLDER"
+            },
+            required:true,
+            pattern: getPattern("Address"),
+           // jsonPath: "Licenses[0].tradeLicenseDetail.owners[0].permanentAddress"
+          }),
       })
-    })
-  });
-};
+    }) }; 
+
+
+
 
 export const transfereeDetails = getCommonCard({
   header: getCommonTitle(
@@ -446,11 +430,11 @@ export const transfereeDetails = getCommonCard({
       applicantType: {
         ...getSelectField({
           label: {
-            labelName: "Applicant Type",
+            labelName: "Ownership Type",
             labelKey: "PT_MUTATION_APPLICANT_TYPE_LABEL"
           },
           placeholder: {
-            labelName: "Select Applicant Type",
+            labelName: "Select Ownership Type",
             labelKey: "PT_MUTATION_APPLICANT_TYPE_LABEL_PLACEHOLDER"
           },
           jsonPath:
@@ -482,73 +466,22 @@ export const transfereeDetails = getCommonCard({
           }
         }),
         beforeFieldChange: (action, state, dispatch) => {
-          let path = action.componentJsonpath.replace(
-            /.applicantType$/,
-            ".applicantSubType"
-          );
-          let applicantType = get(
-            state,
-            "screenConfiguration.preparedFinalObject.applyScreenMdmsData.common-masters.OwnerShipCategory",
-            []
-          );
-          let applicantSubType = applicantType.filter(item => {
-            return item.active && item.code.startsWith(action.value);
-          });
-          dispatch(handleField("apply", path, "props.data", applicantSubType));
-        }
-      },
-      applicantSubType: {
-        ...getSelectField({
-          label: {
-            labelName: "Type of Applicant - Subtype",
-            labelKey: "PT_MUTATION_TRANSFEREE_APPLICANT_SUBTYPE_LABEL"
-          },
-          placeholder: {
-            labelName: "Select Applicant Subtype",
-            labelKey: "PT_MUTATION_APPLICANT_TRANSFEREE_SUBTYPE_PLACEHOLDER"
-          },
-          jsonPath: "FireNOCs[0].fireNOCDetails.applicantDetails.ownerShipType",
-          localePrefix: {
-            moduleName: "common-masters",
-            masterName: "OwnerShipCategory"
-          },
-          // data: [
-          //   {
-          //     code: "Private Company"
-          //   }
-          // ],
-          // props: {
-          //   style: {
-          //     display: "none"
-          //   }
-          // },
-          required: true,
-          gridDefination: {
-            xs: 12,
-            sm: 12,
-            md: 6
-          },
-          props:{
-            className:"applicant-details-error"
-          }
-        }),
-        beforeFieldChange: (action, state, dispatch) => {
+          let path = action.componentJsonpath.replace();
+          // let applicantType = get(
+          //   state,
+          //   "screenConfiguration.preparedFinalObject.applyScreenMdmsData.common-masters.OwnerShipCategory",
+          //   []
+          // );
           let singleApplicantContainerJsonPath =
-            "components.div.children.formwizardThirdStep.children.applicantDetails.children.cardContent.children.applicantTypeContainer.children.singleApplicantContainer";
+          "components.div.children.formwizardFirstStep.children.transfereeDetails.children.cardContent.children.applicantTypeContainer.children.singleApplicantContainer";
           let multipleApplicantContainerJsonPath =
-            "components.div.children.formwizardThirdStep.children.applicantDetails.children.cardContent.children.applicantTypeContainer.children.multipleApplicantContainer";
+          "components.div.children.formwizardFirstStep.children.transfereeDetails.children.cardContent.children.applicantTypeContainer.children.multipleApplicantContainer"
           let institutionContainerJsonPath =
-            "components.div.children.formwizardThirdStep.children.applicantDetails.children.cardContent.children.applicantTypeContainer.children.institutionContainer";
-          // let applicantSubtypeJsonPath =
-          //   "components.div.children.formwizardThirdStep.children.applicantDetails.children.cardContent.children.applicantTypeContainer.children.applicantSubType";
+          "components.div.children.formwizardFirstStep.children.transfereeDetails.children.cardContent.children.applicantTypeContainer.children.institutionContainer";
+         
           if (action.value.includes("SINGLEOWNER")) {
             showComponent(dispatch, singleApplicantContainerJsonPath, true);
             showComponent(dispatch, multipleApplicantContainerJsonPath, false);
-            showComponent(dispatch, institutionContainerJsonPath, false);
-            // showComponent(dispatch, applicantSubtypeJsonPath, false);
-          } else if (action.value.includes("MULTIPLEOWNERS")) {
-            showComponent(dispatch, singleApplicantContainerJsonPath, false);
-            showComponent(dispatch, multipleApplicantContainerJsonPath, true);
             showComponent(dispatch, institutionContainerJsonPath, false);
             // showComponent(dispatch, applicantSubtypeJsonPath, false);
           } else if (action.value.includes("INSTITUTIONAL")) {
@@ -557,8 +490,14 @@ export const transfereeDetails = getCommonCard({
             showComponent(dispatch, institutionContainerJsonPath, true);
             // showComponent(dispatch, applicantSubtypeJsonPath, true);
           }
-        }
-      }
+          else if (action.value.includes("MULTIPLEOWNERS")) {
+            showComponent(dispatch, singleApplicantContainerJsonPath, false);
+            showComponent(dispatch, multipleApplicantContainerJsonPath, true);
+            showComponent(dispatch, institutionContainerJsonPath, false);
+            // showComponent(dispatch, applicantSubtypeJsonPath, true);
+          }
+        },
+      },
     }),
     singleApplicantContainer: {
       uiFramework: "custom-atoms",
